+++ demo_project/config/urls.py
@@ -1,5 +1,5 @@
 """
-URL configuration for demo_project project.
+URL configuration for config project.
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/5.2/topics/http/urls/
@@ -16,7 +16,9 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
+from .views import NPlusOneBookListView
 
 urlpatterns = [
     path('admin/', admin.site.urls),
+    path('books/', NPlusOneBookListView.as_view(), name='n_plus_one_books'),
 ]